Change ship hit information implementation
Because the ship factory is used by the board to create ships, we don't
really have to worry about a ship being created out of bounds.

The ship now creates a `shipInfo` object internally to store both its
"included" coordinates and the hit information when created. This way,
we can easily check the values for sunk and check the keys for inclusion.